packages_requirements.txt
Flask==0.12
mecab-python3==0.7
